Upgrade trigger is needed only for old grub2
Older PLD grub2 packages contained the grub modules in /boot/grub.
After upgrade the installed bootloader would not match the modules
in /boot/grub and fail. When upgrading those old packages to the
new >=2.00 packages the modules in /boot/grub would be completly
missing, causing the same problem.

So the trigger must stay for upgrade from packages <2.00, but is not
needed for future upgrades.